pyproject.toml
[tool.poetry]
name = "rag-box"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.110.1"
python-dotenv = "^1.0.0"
llama-index = "0.10.41"
llama-index-core = "0.10.41"
llama-index-agent-openai = "0.2.6"
uvicorn = { version = "^0.23.2", extras = [ "standard" ] }
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
aiostream = "^0.5.2"
llama-index-vector-stores-chroma = "^0.1.8"
llama-index-embeddings-gemini = "^0.1.6"
llama-index-llms-gemini = "^0.1.10"
llama-index-llms-ollama = "^0.1.3"
llama-index-embeddings-ollama = "^0.1.2"
llama-index-tools-wikipedia = "^0.1.3"
llama-index-tools-duckduckgo = "^0.1.1"
llama-index-vector-stores-qdrant = "^0.2.8"
ollama = "^0.2.0"
llama-index-llms-azure-openai = "^0.1.8"
llama-index-embeddings-azure-openai = "^0.1.9"
cachetools = "^5.3.3"
google-cloud-storage = "^2.17.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
